+++ Riot/Modules/Common/Recents/DataSources/RecentsDataSource.m
@@ -165,10 +165,13 @@ NSString *const kRecentsDataSourceTapOnDirectoryServerChange = @"kRecentsDataSou
 
 - (void)keyBackupStateDidChangeNotification:(NSNotification*)notification
 {
-    [self forceRefresh];
+    if ([self updateKeyBackupBanner])
+    {
+        [self forceRefresh];
+    }
 }
 
-- (void)updateKeyBackupBanner
+- (BOOL)updateKeyBackupBanner
 {
     KeyBackupBanner keyBackupBanner = KeyBackupBannerNone;
     
@@ -207,8 +210,12 @@ NSString *const kRecentsDataSourceTapOnDirectoryServerChange = @"kRecentsDataSou
                 break;
         }
     }
-    
+
+    BOOL updated = (self.keyBackupBanner != keyBackupBanner);
+
     self.keyBackupBanner = keyBackupBanner;
+
+    return updated;
 }
 
 - (void)hideKeyBackupBanner:(KeyBackupBanner)keyBackupBanner
@@ -1227,6 +1234,8 @@ NSString *const kRecentsDataSourceTapOnDirectoryServerChange = @"kRecentsDataSou
 
 - (void)refreshRoomsSections
 {
+    NSDate *startDate = [NSDate date];
+
     [invitesCellDataArray removeAllObjects];
     [favoriteCellDataArray removeAllObjects];
     [peopleCellDataArray removeAllObjects];
@@ -1239,8 +1248,6 @@ NSString *const kRecentsDataSourceTapOnDirectoryServerChange = @"kRecentsDataSou
     _missedGroupDiscussionsCount = _missedHighlightGroupDiscussionsCount = 0;
     
     keyBackupBannerSection = directorySection = favoritesSection = peopleSection = conversationSection = lowPrioritySection = serverNoticeSection = invitesSection = -1;
-    
-    [self updateKeyBackupBanner];
     
     if (displayedRecentsDataSourceArray.count > 0)
     {
@@ -1485,6 +1492,8 @@ NSString *const kRecentsDataSourceTapOnDirectoryServerChange = @"kRecentsDataSou
             }];
         }
     }
+
+    NSLog(@"[RecentsDataSource] refreshRoomsSections: Done in %.0fms", [[NSDate date] timeIntervalSinceDate:startDate] * 1000);
 }
 
 - (void)dataSource:(MXKDataSource*)dataSource didCellChange:(id)changes