+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes in 0.8.1 (2019-02-)
 Bug fix:
  * Fix text color in room preview (PR #2261)
  * Fix navigation bar background after accepting an invite (PR #2261)
+ * Tabs at the top of Room Details are hard to see in dark theme (#2260).
 
 Changes in 0.8.0 (2019-02-15)
 ===============================================